requirements.txt
tensorflow-gpu >= 1.3.0
setproctitle
redis
pyyaml